requirements.txt
pytorch==1.7.0
torchvision==0.8.1
pytorch-transformers==1.1.0
numpy==1.16.4
lmdb==0.94
tensorflow==1.15.2
tensorpack==0.9.4
tqdm==4.31.1
easydict==1.9
PyYAML==5.4
jsonlines==1.2.0
json-lines==0.5.0
matplotlib
Cython==0.29.21
python-prctl
msgpack
msgpack-numpy
opencv-python==4.2.0.34
clip==1.0
pandas==1.1.5
lmdb==0.94
easydict==1.9
boto3==1.16.13
botocore==1.19.13
seaborn==0.11.1
scipy==1.5.4
scikit-learn==0.24.2